update to 7.7

+++ specification/specs/document/single/term_vectors/FieldStatistics.ts
@@ -1,5 +1,5 @@
 class FieldStatistics {
 	doc_count: integer;
-	sum_doc_freq: integer;
-	sum_ttf: integer;
+	sum_doc_freq: long;
+	sum_ttf: long;
 }
